+++ src/Common/StreamTools.cpp
@@ -1,6 +1,19 @@
-// Copyright (c) 2011-2016 The Cryptonote developers
-// Distributed under the MIT/X11 software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+// Copyright (c) 2012-2017, The CryptoNote developers, The Bytecoin developers
+//
+// This file is part of Bytecoin.
+//
+// Bytecoin is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Bytecoin is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with Bytecoin.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "StreamTools.h"
 #include <stdexcept>
@@ -9,9 +22,9 @@
 
 namespace Common {
 
-void read(IInputStream& in, void* data, size_t size) {
+void read(IInputStream& in, void* data, uint64_t size) {
   while (size > 0) {
-    size_t readSize = in.readSome(data, size);
+    uint64_t readSize = in.readSome(data, size);
     if (readSize == 0) {
       throw std::runtime_error("Failed to read from IInputStream");
     }
@@ -59,12 +72,12 @@ void read(IInputStream& in, uint64_t& value) {
   read(in, &value, sizeof(value));
 }
 
-void read(IInputStream& in, std::vector<uint8_t>& data, size_t size) {
+void read(IInputStream& in, std::vector<uint8_t>& data, uint64_t size) {
   data.resize(size);
   read(in, data.data(), size);
 }
 
-void read(IInputStream& in, std::string& data, size_t size) {
+void read(IInputStream& in, std::string& data, uint64_t size) {
   std::vector<char> temp(size);
   read(in, temp.data(), size);
   data.assign(temp.data(), size);
@@ -79,7 +92,7 @@ void readVarint(IInputStream& in, uint8_t& value) {
       throw std::runtime_error("readVarint, value overflow");
     }
 
-    temp |= static_cast<size_t>(piece & 0x7f) << shift;
+    temp |= static_cast<uint64_t>(piece & 0x7f) << shift;
     if ((piece & 0x80) == 0) {
       if (piece == 0 && shift != 0) {
         throw std::runtime_error("readVarint, invalid value representation");
@@ -101,7 +114,7 @@ void readVarint(IInputStream& in, uint16_t& value) {
       throw std::runtime_error("readVarint, value overflow");
     }
 
-    temp |= static_cast<size_t>(piece & 0x7f) << shift;
+    temp |= static_cast<uint64_t>(piece & 0x7f) << shift;
     if ((piece & 0x80) == 0) {
       if (piece == 0 && shift != 0) {
         throw std::runtime_error("readVarint, invalid value representation");
@@ -123,7 +136,7 @@ void readVarint(IInputStream& in, uint32_t& value) {
       throw std::runtime_error("readVarint, value overflow");
     }
 
-    temp |= static_cast<size_t>(piece & 0x7f) << shift;
+    temp |= static_cast<uint64_t>(piece & 0x7f) << shift;
     if ((piece & 0x80) == 0) {
       if (piece == 0 && shift != 0) {
         throw std::runtime_error("readVarint, invalid value representation");
@@ -158,9 +171,9 @@ void readVarint(IInputStream& in, uint64_t& value) {
   value = temp;
 }
 
-void write(IOutputStream& out, const void* data, size_t size) {
+void write(IOutputStream& out, const void* data, uint64_t size) {
   while (size > 0) {
-    size_t writtenSize = out.writeSome(data, size);
+    uint64_t writtenSize = out.writeSome(data, size);
     if (writtenSize == 0) {
       throw std::runtime_error("Failed to write to IOutputStream");
     }
