+++ src/Common/StringOutputStream.cpp
@@ -1,6 +1,19 @@
-// Copyright (c) 2011-2016 The Cryptonote developers
-// Distributed under the MIT/X11 software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+// Copyright (c) 2012-2017, The CryptoNote developers, The Bytecoin developers
+//
+// This file is part of Bytecoin.
+//
+// Bytecoin is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Bytecoin is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with Bytecoin.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "StringOutputStream.h"
 
@@ -9,7 +22,7 @@ namespace Common {
 StringOutputStream::StringOutputStream(std::string& out) : out(out) {
 }
 
-size_t StringOutputStream::writeSome(const void* data, size_t size) {
+uint64_t StringOutputStream::writeSome(const void* data, uint64_t size) {
   out.append(static_cast<const char*>(data), size);
   return size;
 }
