+++ src/Serialization/BinaryOutputStreamSerializer.cpp
@@ -1,6 +1,19 @@
-// Copyright (c) 2011-2016 The Cryptonote developers
-// Distributed under the MIT/X11 software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+// Copyright (c) 2012-2017, The CryptoNote developers, The Bytecoin developers
+//
+// This file is part of Bytecoin.
+//
+// Bytecoin is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Bytecoin is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with Bytecoin.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "BinaryOutputStreamSerializer.h"
 
@@ -23,7 +36,7 @@ bool BinaryOutputStreamSerializer::beginObject(Common::StringView name) {
 void BinaryOutputStreamSerializer::endObject() {
 }
 
-bool BinaryOutputStreamSerializer::beginArray(size_t& size, Common::StringView name) {
+bool BinaryOutputStreamSerializer::beginArray(uint64_t& size, Common::StringView name) {
   writeVarint(stream, size);
   return true;
 }
@@ -78,7 +91,7 @@ bool BinaryOutputStreamSerializer::operator()(std::string& value, Common::String
   return true;
 }
 
-bool BinaryOutputStreamSerializer::binary(void* value, size_t size, Common::StringView name) {
+bool BinaryOutputStreamSerializer::binary(void* value, uint64_t size, Common::StringView name) {
   checkedWrite(static_cast<const char*>(value), size);
   return true;
 }
@@ -94,7 +107,7 @@ bool BinaryOutputStreamSerializer::operator()(double& value, Common::StringView 
   return false;
 }
 
-void BinaryOutputStreamSerializer::checkedWrite(const char* buf, size_t size) {
+void BinaryOutputStreamSerializer::checkedWrite(const char* buf, uint64_t size) {
   write(stream, buf, size);
 }
