+++ src/Serialization/JsonInputValueSerializer.cpp
@@ -1,6 +1,19 @@
-// Copyright (c) 2011-2016 The Cryptonote developers
-// Distributed under the MIT/X11 software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+// Copyright (c) 2012-2017, The CryptoNote developers, The Bytecoin developers
+//
+// This file is part of Bytecoin.
+//
+// Bytecoin is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Bytecoin is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with Bytecoin.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "JsonInputValueSerializer.h"
 
@@ -58,7 +71,7 @@ void JsonInputValueSerializer::endObject() {
   chain.pop_back();
 }
 
-bool JsonInputValueSerializer::beginArray(size_t& size, Common::StringView name) {
+bool JsonInputValueSerializer::beginArray(uint64_t& size, Common::StringView name) {
   const JsonValue* parent = chain.back();
   std::string strName(name);
 
@@ -132,7 +145,7 @@ bool JsonInputValueSerializer::operator()(bool& value, Common::StringView name) 
   return true;
 }
 
-bool JsonInputValueSerializer::binary(void* value, size_t size, Common::StringView name) {
+bool JsonInputValueSerializer::binary(void* value, uint64_t size, Common::StringView name) {
   auto ptr = getValue(name);
   if (ptr == nullptr) {
     return false;
