+++ src/Serialization/JsonOutputStreamSerializer.cpp
@@ -1,6 +1,19 @@
-// Copyright (c) 2011-2016 The Cryptonote developers
-// Distributed under the MIT/X11 software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+// Copyright (c) 2012-2017, The CryptoNote developers, The Bytecoin developers
+//
+// This file is part of Bytecoin.
+//
+// Bytecoin is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Bytecoin is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with Bytecoin.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "JsonOutputStreamSerializer.h"
 #include <cassert>
@@ -59,7 +72,7 @@ void JsonOutputStreamSerializer::endObject() {
   chain.pop_back();
 }
 
-bool JsonOutputStreamSerializer::beginArray(size_t& size, Common::StringView name) {
+bool JsonOutputStreamSerializer::beginArray(uint64_t& size, Common::StringView name) {
   JsonValue val(JsonValue::ARRAY);
   JsonValue& res = chain.back()->insert(std::string(name), val);
   chain.push_back(&res);
@@ -121,7 +134,7 @@ bool JsonOutputStreamSerializer::operator()(bool& value, Common::StringView name
   return true;
 }
 
-bool JsonOutputStreamSerializer::binary(void* value, size_t size, Common::StringView name) {
+bool JsonOutputStreamSerializer::binary(void* value, uint64_t size, Common::StringView name) {
   std::string hex = Common::toHex(value, size);
   return (*this)(hex, name);
 }
