+++ src/Serialization/KVBinaryOutputStreamSerializer.h
@@ -1,6 +1,19 @@
-// Copyright (c) 2011-2016 The Cryptonote developers
-// Distributed under the MIT/X11 software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+// Copyright (c) 2012-2017, The CryptoNote developers, The Bytecoin developers
+//
+// This file is part of Bytecoin.
+//
+// Bytecoin is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Bytecoin is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with Bytecoin.  If not, see <http://www.gnu.org/licenses/>.
 
 #pragma once
 
@@ -24,7 +37,7 @@ public:
   virtual bool beginObject(Common::StringView name) override;
   virtual void endObject() override;
 
-  virtual bool beginArray(size_t& size, Common::StringView name) override;
+  virtual bool beginArray(uint64_t& size, Common::StringView name) override;
   virtual void endArray() override;
 
   virtual bool operator()(uint8_t& value, Common::StringView name) override;
@@ -37,7 +50,7 @@ public:
   virtual bool operator()(double& value, Common::StringView name) override;
   virtual bool operator()(bool& value, Common::StringView name) override;
   virtual bool operator()(std::string& value, Common::StringView name) override;
-  virtual bool binary(void* value, size_t size, Common::StringView name) override;
+  virtual bool binary(void* value, uint64_t size, Common::StringView name) override;
   virtual bool binary(std::string& value, Common::StringView name) override;
 
   template<typename T>
@@ -62,12 +75,12 @@ private:
   struct Level {
     State state;
     std::string name;
-    size_t count;
+    uint64_t count;
 
     Level(Common::StringView nm) :
       name(nm), state(State::Object), count(0) {}
 
-    Level(Common::StringView nm, size_t arraySize) :
+    Level(Common::StringView nm, uint64_t arraySize) :
       name(nm), state(State::ArrayPrefix), count(arraySize) {}
 
     Level(Level&& rv) {
