+++ src/Serialization/MemoryStream.h
@@ -1,6 +1,19 @@
-// Copyright (c) 2011-2016 The Cryptonote developers
-// Distributed under the MIT/X11 software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+// Copyright (c) 2012-2017, The CryptoNote developers, The Bytecoin developers
+//
+// This file is part of Bytecoin.
+//
+// Bytecoin is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Bytecoin is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with Bytecoin.  If not, see <http://www.gnu.org/licenses/>.
 
 #pragma once
 
@@ -18,7 +31,7 @@ public:
   MemoryStream() : m_writePos(0) {
   }
 
-  virtual size_t writeSome(const void* data, size_t size) override {
+  virtual uint64_t writeSome(const void* data, uint64_t size) override {
     if (size == 0) {
       return 0;
     }
@@ -32,7 +45,7 @@ public:
     return size;
   }
 
-  size_t size() {
+  uint64_t size() {
     return m_buffer.size();
   }
 
@@ -46,7 +59,7 @@ public:
   }
 
 private:
-  size_t m_writePos;
+  uint64_t m_writePos;
   std::vector<uint8_t> m_buffer;
 };
